Improve Unit Tests
